Revert "refactor: change directory structure"
This reverts commit c76a83f6ce8262fd15cd3632bd9a10130607991a.
