--- src/MapView/MapView/index.ts
@@ -1,4 +1,0 @@
-import MapView from './MapView';
-
-export type {MapViewProps, MapViewHandle} from './MapViewTypes';
-export default MapView;
