+++ src/MapView/index.ts
@@ -1,4 +1,4 @@
 import MapView from './MapView';
 
-export * from './MapView';
+export type {MapViewProps, MapViewHandle} from './MapViewTypes';
 export default MapView;
